Moved two new factors into new files and renamed test
