+++ Tests/SwiftFusionTests/Inference/SwitchingMCMCTests.swift
@@ -1,100 +1,16 @@
+//
+//  SwitchingMCMCTests.swift
+//
+//
+//  Frank Dellaert and Marc Rasi
+//  July 2020
+
 import Foundation
 import TensorFlow
 import XCTest
 
 import PenguinStructures
 @testable import SwiftFusion
-
-/// A factor on two discrete labels evaluation the transition probability
-struct DiscreteTransitionFactor : Factor {
-  typealias Variables = Tuple2<Int, Int>
-  
-  /// The IDs of the variables adjacent to this factor.
-  public let edges: Variables.Indices
-  
-  /// The number of states.
-  let stateCount: Int
-  
-  /// Entry `i * stateCount + j` is the probability of transitioning from state `j` to state `i`.
-  let transitionMatrix: [Double]
-  
-  init(
-    _ inputId1: TypedID<Int, Int>,
-    _ inputId2: TypedID<Int, Int>,
-    _ stateCount: Int,
-    _ transitionMatrix: [Double]
-  ) {
-    precondition(transitionMatrix.count == stateCount * stateCount)
-    self.edges = Tuple2(inputId1, inputId2)
-    self.stateCount = stateCount
-    self.transitionMatrix = transitionMatrix
-  }
-  
-  func error(at q: Variables) -> Double {
-    let (label1, label2) = (q.head, q.tail.head)
-    return -log(transitionMatrix[label2 * stateCount + label1])
-  }
-}
-
-/// A factor with a switchable motion model.
-///
-/// `JacobianRows` specifies the `Rows` parameter of the Jacobian of this factor. See the
-/// documentation on `JacobianFactor.jacobian` for more information. Use the typealiases below to
-/// avoid specifying this type parameter every time you create an instance.
-public struct SwitchingBetweenFactor<Pose: LieGroup, JacobianRows: FixedSizeArray>:
-  LinearizableFactor
-where JacobianRows.Element == Tuple2<Pose.TangentVector, Pose.TangentVector>
-{
-  public typealias Variables = Tuple3<Pose, Int, Pose>
-  
-  public let edges: Variables.Indices
-  
-  /// Movement temmplates for each label.
-  let movements: [Pose]
-  
-  public init(_ from: TypedID<Pose, Int>,
-              _ label: TypedID<Int, Int>,
-              _ to: TypedID<Pose, Int>,
-              _ movements: [Pose]) {
-    self.edges = Tuple3(from, label, to)
-    self.movements = movements
-  }
-  
-  public typealias ErrorVector = Pose.TangentVector
-  
-  @differentiable(wrt: (start, end))
-  public func errorVector(_ start: Pose, _ label: Int, _ end: Pose) -> ErrorVector {
-    let actualMotion = between(start, end)
-    return movements[label].localCoordinate(actualMotion)
-  }
-  
-  // Note: All the remaining code in this factor is boilerplate that we can eventually eliminate
-  // with sugar.
-  
-  public func error(at x: Variables) -> Double {
-    return errorVector(at: x).squaredNorm
-  }
-  
-  public func errorVector(at x: Variables) -> Pose.TangentVector {
-    return errorVector(x.head, x.tail.head, x.tail.tail.head)
-  }
-  
-  public typealias Linearization = JacobianFactor<JacobianRows, ErrorVector>
-  public func linearized(at x: Variables) -> Linearization {
-    let (start, label, end) = (x.head, x.tail.head, x.tail.tail.head)
-    let (startEdge, endEdge) = (edges.head, edges.tail.tail.head)
-    let differentiableVariables = Tuple2(start, end)
-    let differentiableEdges = Tuple2(startEdge, endEdge)
-    return Linearization(
-      linearizing: { differentiableVariables in
-        let (start, end) = (differentiableVariables.head, differentiableVariables.tail.head)
-        return errorVector(start, label, end)
-      },
-      at: differentiableVariables,
-      edges: differentiableEdges
-    )
-  }
-}
 
 /// A between factor on `Pose2`.
 public typealias SwitchingBetweenFactor2 = SwitchingBetweenFactor<Pose2, Array3<Tuple2<Vector3, Vector3>>>
@@ -121,20 +37,20 @@ class Scratch: XCTestCase {
     let x3 = variables.store(origin)
     
     var graph = FactorGraph()
-    graph.store(PriorFactor2(x1, z1)) // prior
-    graph.store(PriorFactor2(x1, z1))
-    graph.store(PriorFactor2(x2, z2))
-    graph.store(PriorFactor2(x3, z3))
-    graph.store(BetweenFactor2(x1, x2, forwardMove))
-    graph.store(BetweenFactor2(x2, x3, forwardMove))
+    graph.store(PriorFactor(x1, z1)) // prior
+    graph.store(PriorFactor(x1, z1))
+    graph.store(PriorFactor(x2, z2))
+    graph.store(PriorFactor(x3, z3))
+    graph.store(BetweenFactor(x1, x2, forwardMove))
+    graph.store(BetweenFactor(x2, x3, forwardMove))
     
     return (graph, variables)
   }
   
   func printPoses(_ variables : VariableAssignments) {
-    print(variables[TypedID<Pose2, Int>(0)])
-    print(variables[TypedID<Pose2, Int>(1)])
-    print(variables[TypedID<Pose2, Int>(2)])
+    print(variables[TypedID<Pose2>(0)])
+    print(variables[TypedID<Pose2>(1)])
+    print(variables[TypedID<Pose2>(2)])
   }
   
   /// Tracking example from Figure 2.a
@@ -179,10 +95,10 @@ class Scratch: XCTestCase {
     ]
     
     var graph = FactorGraph()
-    graph.store(PriorFactor2(x1, z1)) // prior
-    graph.store(PriorFactor2(x1, z1))
-    graph.store(PriorFactor2(x2, z2))
-    graph.store(PriorFactor2(x3, z3))
+    graph.store(PriorFactor(x1, z1)) // prior
+    graph.store(PriorFactor(x1, z1))
+    graph.store(PriorFactor(x2, z2))
+    graph.store(PriorFactor(x3, z3))
     graph.store(SwitchingBetweenFactor2(x1, q1, x2, movements))
     graph.store(SwitchingBetweenFactor2(x2, q2, x3, movements))
     graph.store(DiscreteTransitionFactor(q1, q2, labelCount, transitionMatrix))
@@ -191,8 +107,8 @@ class Scratch: XCTestCase {
   }
   
   func printLabels(_ variables : VariableAssignments) {
-    print(variables[TypedID<Int, Int>(0)])
-    print(variables[TypedID<Int, Int>(1)])
+    print(variables[TypedID<Int>(0)])
+    print(variables[TypedID<Int>(1)])
   }
 
   /// Returns the gradient of the error function of `graph` at `x`.
@@ -203,7 +119,7 @@ class Scratch: XCTestCase {
 
   func printGradient(_ grad: AllVectors) {
     for i in 0..<3 {
-      print("  \(grad[TypedID<Vector3, Int>(i)])")
+      print("  \(grad[TypedID<Vector3>(i)])")
     }
   }
   
@@ -239,11 +155,11 @@ class Scratch: XCTestCase {
     /// Proposal to change one label, and re-optimize
     let flipAndOptimize = {(x:VariableAssignments) -> VariableAssignments in
       let labelVars = x.storage[ObjectIdentifier(Int.self)]
-      let positionVars = x.storage[ObjectIdentifier(Pose2.self)]
+//      let positionVars = x.storage[ObjectIdentifier(Pose2.self)]
       
       // Randomly change one label.
       let i = Int.random(in: 0..<labelVars!.count)
-      let id = TypedID<Int, Int>(i)
+      let id = TypedID<Int>(i)
       var y = x
       y[id] = Int.random(in: 0..<3)
